pyproject.toml
[tool.poetry]
name = "fast-api-mongo-template"
version = "0.1.0"
description = "Fast API mongo db starter template"
authors = ["Milos Petrovic <milos.brugi@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_api_mongo_template"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.81.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
motor = "^3.0.0"
dnspython = "^2.2.1"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
requests = "^2.28.1"
coverage = "^6.4.4"
pytest-mock = "^3.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
